Implemented col of numbers levels.
Need to rethink the order of levels. Difficulty jumps around too much.

Also need to make progress "stickier". The first few errors should not count for
much, but further errors should grow in importance. Basically if you get a
constant 1 in N correct you should progress (faster if N is smaller). As is you
need to get more than half of the answers correct or you regress.

I think the trick is to push forward and let a few errors go, but if the player
starts to make a lot of mistakes, reduce the level until they are getting some
portion correct on the first shot.